pyproject.toml
[project]
name = "hospital-fastapi"
version = "0.1.0"
description = ""
authors = [
    {name = "Mithilesh",email = "mithileshchaurasiya1627@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio (>=4.0.0,<5.0.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "motor (>=3.7.1,<4.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "bcrypt (==3.2.2)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
